+++ assignments/assignment3/banking/src/main/java/dk/cphbusiness/banking/Customer.java
@@ -3,7 +3,7 @@ package dk.cphbusiness.banking;
 import java.util.List;
 
 public interface Customer {
-    void transfer(long amount, Account account, Customer target);
+    void transfer(double amount, Account account, Customer target);
     String getCpr();
     List<Account> getAccounts();
     List<Movement> getMovementsFromAccount(String accountNumber);
