+++ assignments/assignment3/banking/src/test/java/dk/cphbusiness/banking/CustomerStub.java
@@ -11,7 +11,7 @@ public class CustomerStub implements Customer {
     }
 
     @Override
-    public void transfer(long amount, Account account, Customer target) {
+    public void transfer(double amount, Account account, Customer target) {
     }
 
     @Override
